Eliminate JoinUtil by moving methods to BaseOperation class
