+++ ksml-runner/src/main/java/io/axual/ksml/runner/backend/KafkaStreamsRunner.java
@@ -26,8 +26,8 @@ import io.axual.ksml.client.resolving.ResolvingClientConfig;
 import io.axual.ksml.execution.ExecutionContext;
 import io.axual.ksml.execution.ExecutionErrorHandler;
 import io.axual.ksml.generator.TopologyDefinition;
-import io.axual.ksml.metric.KsmlTagEnricher;
 import io.axual.ksml.metric.KsmlMetricsReporter;
+import io.axual.ksml.metric.KsmlTagEnricher;
 import io.axual.ksml.python.PythonContextConfig;
 import io.axual.ksml.runner.config.ApplicationServerConfig;
 import io.axual.ksml.runner.exception.RunnerException;
