+++ ksml/src/main/java/io/axual/ksml/metric/KsmlMetricsReporter.java
@@ -20,20 +20,15 @@ package io.axual.ksml.metric;
  * =========================LICENSE_END==================================
  */
 
+import lombok.NoArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.common.metrics.KafkaMetric;
 import org.apache.kafka.common.metrics.MetricsReporter;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.function.DoubleSupplier;
-
-import lombok.NoArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 /**
  * A Kafka MetricsReporter implementation that converts selected Kafka Streams metrics
