+++ ksml/src/main/java/io/axual/ksml/operation/GroupByKeyOperation.java
@@ -22,7 +22,6 @@ package io.axual.ksml.operation;
 
 
 import io.axual.ksml.generator.TopologyBuildContext;
-import io.axual.ksml.store.StoreUtil;
 import io.axual.ksml.stream.KGroupedStreamWrapper;
 import io.axual.ksml.stream.KStreamWrapper;
 import io.axual.ksml.stream.StreamWrapper;
@@ -42,7 +41,7 @@ public class GroupByKeyOperation extends StoreOperation {
 
         final var k = input.keyType();
         final var v = input.valueType();
-        final var kvStore = StoreUtil.validateKeyValueStore(this, store(), k, v);
+        final var kvStore = validateKeyValueStore(store(), k, v);
         final var grouped = groupedOf(k, v, kvStore);
         final KGroupedStream<Object, Object> output = grouped != null
                 ? input.stream.groupByKey(grouped)
