+++ ksml/src/main/java/io/axual/ksml/operation/LeftJoinWithGlobalTableOperation.java
@@ -27,7 +27,6 @@ import io.axual.ksml.generator.TopologyBuildContext;
 import io.axual.ksml.stream.KStreamWrapper;
 import io.axual.ksml.stream.StreamWrapper;
 import io.axual.ksml.user.UserKeyTransformer;
-import io.axual.ksml.util.JoinUtil;
 import org.apache.kafka.streams.kstream.KStream;
 
 public class LeftJoinWithGlobalTableOperation extends BaseOperation {
@@ -66,7 +65,7 @@ public class LeftJoinWithGlobalTableOperation extends BaseOperation {
         final var sel = userFunctionOf(context, KEYSELECTOR_NAME, keySelector, subOf(gk), superOf(k), superOf(v));
         final var joiner = userFunctionOf(context, VALUEJOINER_NAME, valueJoiner, subOf(rv), superOf(k), superOf(v), superOf(gv));
         final var userSel = new UserKeyTransformer(sel, tags);
-        final var userJoiner = JoinUtil.valueJoinerWithKey(joiner, tags);
+        final var userJoiner = valueJoinerWithKey(joiner, tags);
         final var named = namedOf();
         final KStream<Object, Object> output = named != null
                 ? input.stream.leftJoin(otherGlobalKTable.globalTable, userSel, userJoiner, named)
