+++ ksml/src/main/java/io/axual/ksml/operation/OuterJoinWithStreamOperation.java
@@ -24,10 +24,8 @@ package io.axual.ksml.operation;
 import io.axual.ksml.definition.FunctionDefinition;
 import io.axual.ksml.definition.StreamDefinition;
 import io.axual.ksml.generator.TopologyBuildContext;
-import io.axual.ksml.store.StoreUtil;
 import io.axual.ksml.stream.KStreamWrapper;
 import io.axual.ksml.stream.StreamWrapper;
-import io.axual.ksml.util.JoinUtil;
 import org.apache.kafka.streams.kstream.JoinWindows;
 import org.apache.kafka.streams.kstream.KStream;
 
@@ -43,7 +41,7 @@ public class OuterJoinWithStreamOperation extends DualStoreOperation {
         super(config);
         this.joinStream = joinStream;
         this.valueJoiner = valueJoiner;
-        this.joinWindows = JoinUtil.joinWindowsOf(timeDifference, gracePeriod);
+        this.joinWindows = joinWindowsOf(timeDifference, gracePeriod);
     }
 
     @Override
@@ -65,10 +63,10 @@ public class OuterJoinWithStreamOperation extends DualStoreOperation {
         final var vr = streamDataTypeOf(firstSpecificType(valueJoiner, vo, v), false);
         checkType("Join stream keyType", ko, equalTo(k));
         final var joiner = userFunctionOf(context, VALUEJOINER_NAME, valueJoiner, subOf(vr), superOf(k), superOf(v), superOf(vo));
-        final var windowStore1 = StoreUtil.validateWindowStore(this, store1(), k, vr);
-        final var windowStore2 = StoreUtil.validateWindowStore(this, store2(), k, vr);
-        final var streamJoined = JoinUtil.streamJoinedOf(name, windowStore1, windowStore2, k, v, vo);
-        final var userJoiner = JoinUtil.valueJoiner(joiner, tags);
+        final var thisStore = validateWindowStore(thisStore(), k, vr);
+        final var otherStore = validateWindowStore(otherStore(), k, vr);
+        final var streamJoined = streamJoinedOf(thisStore, otherStore, k, v, vo);
+        final var userJoiner = valueJoiner(joiner, tags);
         final KStream<Object, Object> output = streamJoined != null
                 ? input.stream.outerJoin(otherStream.stream, userJoiner, joinWindows, streamJoined)
                 : input.stream.outerJoin(otherStream.stream, userJoiner, joinWindows);
