+++ ksml/src/main/java/io/axual/ksml/store/StoreUtil.java
@@ -24,10 +24,7 @@ import io.axual.ksml.definition.KeyValueStateStoreDefinition;
 import io.axual.ksml.definition.SessionStateStoreDefinition;
 import io.axual.ksml.definition.StateStoreDefinition;
 import io.axual.ksml.definition.WindowStateStoreDefinition;
-import io.axual.ksml.exception.ExecutionException;
 import io.axual.ksml.generator.StreamDataType;
-import io.axual.ksml.operation.BaseOperation;
-import io.axual.ksml.type.UserType;
 import org.apache.kafka.common.serialization.Serde;
 import org.apache.kafka.common.utils.Bytes;
 import org.apache.kafka.streams.kstream.Materialized;
@@ -142,96 +139,5 @@ public class StoreUtil {
         mat = store.caching() ? mat.withCachingEnabled() : mat.withCachingDisabled();
         mat = store.logging() ? mat.withLoggingEnabled(new HashMap<>()) : mat.withLoggingDisabled();
         return new MaterializedStore<>(mat, keySerde, valueSerde);
-    }
-
-    public static KeyValueStateStoreDefinition validateKeyValueStore(BaseOperation operation, StateStoreDefinition store, StreamDataType keyType, StreamDataType valueType) {
-        return validateKeyValueStore(operation, store, keyType.userType(), valueType.userType());
-    }
-
-    public static KeyValueStateStoreDefinition validateKeyValueStore(BaseOperation operation, StateStoreDefinition store, UserType keyType, UserType valueType) {
-        if (store == null) return null;
-        if (store instanceof KeyValueStateStoreDefinition keyValueStore) {
-            validateStore(store, keyType, valueType);
-            final var storeKeyType = keyType != null ? keyType : keyValueStore.keyType();
-            final var storeValueType = valueType != null ? valueType : keyValueStore.valueType();
-            return new KeyValueStateStoreDefinition(
-                    keyValueStore.name(),
-                    keyValueStore.persistent(),
-                    keyValueStore.timestamped(),
-                    keyValueStore.versioned(),
-                    keyValueStore.historyRetention(),
-                    keyValueStore.segmentInterval(),
-                    storeKeyType,
-                    storeValueType,
-                    keyValueStore.caching(),
-                    keyValueStore.logging());
-        }
-        throw new ExecutionException(operation + " requires a  state store of type 'keyValue'");
-    }
-
-    public static SessionStateStoreDefinition validateSessionStore(BaseOperation operation, StateStoreDefinition store, StreamDataType keyType, StreamDataType valueType) {
-        return validateSessionStore(operation, store, keyType.userType(), valueType.userType());
-    }
-
-    public static SessionStateStoreDefinition validateSessionStore(BaseOperation operation, StateStoreDefinition store, UserType keyType, UserType valueType) {
-        if (store == null) return null;
-        if (store instanceof SessionStateStoreDefinition sessionStore) {
-            validateStore(store, keyType, valueType);
-            final var storeKeyType = keyType != null ? keyType : sessionStore.keyType();
-            final var storeValueType = valueType != null ? valueType : sessionStore.valueType();
-            return new SessionStateStoreDefinition(
-                    sessionStore.name(),
-                    sessionStore.persistent(),
-                    sessionStore.timestamped(),
-                    sessionStore.retention(),
-                    storeKeyType,
-                    storeValueType,
-                    sessionStore.caching(),
-                    sessionStore.logging());
-        }
-        throw new ExecutionException(operation + " requires a  state store of type 'session'");
-    }
-
-    public static WindowStateStoreDefinition validateWindowStore(BaseOperation operation, StateStoreDefinition store, StreamDataType keyType, StreamDataType valueType) {
-        return validateWindowStore(operation, store, keyType.userType(), valueType.userType());
-    }
-
-    public static WindowStateStoreDefinition validateWindowStore(BaseOperation operation, StateStoreDefinition store, UserType keyType, UserType valueType) {
-        if (store == null) return null;
-        if (store instanceof WindowStateStoreDefinition windowStore) {
-            validateStore(store, keyType, valueType);
-            final var storeKeyType = keyType != null ? keyType : windowStore.keyType();
-            final var storeValueType = valueType != null ? valueType : windowStore.valueType();
-            return new WindowStateStoreDefinition(
-                    windowStore.name(),
-                    windowStore.persistent(),
-                    windowStore.timestamped(),
-                    windowStore.retention(),
-                    windowStore.windowSize(),
-                    windowStore.retainDuplicates(),
-                    storeKeyType,
-                    storeValueType,
-                    windowStore.caching(),
-                    windowStore.logging());
-        }
-        throw new ExecutionException(operation + " requires a  state store of type 'window'");
-    }
-
-    private static void validateStore(StateStoreDefinition store, UserType keyType, UserType valueType) {
-        validateStoreTypeWithStreamType(store, "key", store.keyType(), keyType);
-        validateStoreTypeWithStreamType(store, "value", store.valueType(), valueType);
-    }
-
-    private static void validateStoreTypeWithStreamType(StateStoreDefinition store, String keyOrValue, UserType storeKeyOrValueType, UserType streamKeyOrValueType) {
-        if (streamKeyOrValueType == null) {
-            if (storeKeyOrValueType == null) {
-                throw new ExecutionException("State store '" + store.name() + "' does not have a defined " + keyOrValue + " type");
-            }
-            return;
-        }
-
-        if (storeKeyOrValueType != null && !storeKeyOrValueType.dataType().isAssignableFrom(streamKeyOrValueType.dataType())) {
-            throw new ExecutionException("Incompatible " + keyOrValue + " types for state store '" + store.name() + "': " + storeKeyOrValueType + " and " + streamKeyOrValueType);
-        }
     }
 }
